Move PWSLocale class def to a new file
