+++ src/ui/wxWidgets/PWSafeApp.h
@@ -26,7 +26,8 @@
 #include "SafeCombinationEntryDlg.h"
 ////@end includes
 #include "core/PWScore.h"
-#include "./RecentDbList.h"
+#include "RecentDbList.h"
+#include "PWSLocale.h"
 
 /*!
  * Forward declarations
@@ -43,42 +44,6 @@ class PasswordSafeFrame;
 
 ////@begin control identifiers
 ////@end control identifiers
-
-// wx3.2.2 and later can use wxUILocale and some improved logic
-#if wxCHECK_VERSION(3, 2, 2)
-#define LOCALE_WX322
-#endif
-
-class PWSMacLocale
-{
-public:
-#ifdef __WXMAC__
-  static void setMacLocale(const char *loc);
-#else
-  static void setMacLocale([[maybe_unused]] const char *loc) {}; // no-op if not macOS
-#endif
-};
-
-#ifdef LOCALE_WX322
-#include <wx/uilocale.h>
-
-  class PWSLocale : public wxUILocale, public PWSMacLocale
-  {
-  public:
-    static wxString PWSGetCurrentName() { return wxUILocale::GetCurrent().GetName(); };
-  };
-
-#else // wxCHECK_VERSION
-
-  // Pre-wx3.2.2 compatible version
-  class PWSLocale : public wxLocale, public PWSMacLocale
-  {
-  public:
-    PWSLocale() {};
-    static bool UseDefault() { return false; }; //no-op for compatibility
-    wxString PWSGetCurrentName() { return wxLocale::GetCanonicalName(); };
-  };
-#endif // wxCHECK_VERSION
 
 /*!
  * PWSafeApp class declaration
